requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.1
matplotlib==3.7.1
seaborn==0.12.2

# Machine learning dependencies
scikit-learn==1.2.2
joblib==1.2.0

# Natural language processing dependencies
transformers==4.28.1
langchain==0.0.168
pytorch==2.0.1
sentencepiece==0.1.99
faiss-cpu==1.7.4

# API and integration
requests==2.29.0
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8

# Testing and development
pytest==7.3.1
black==23.3.0 